+++ code/src/main/java/com/project/accountbook/card/repository/CardDAO.java
@@ -4,9 +4,8 @@ import java.sql.Connection;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 import java.sql.Statement;
-import java.util.HashMap;
 
-import com.project.accountbook.card.model.MyCardDTO;
+import com.project.accountbook.card.model.CardDTO;
 import com.project.accountbook.util.DBUtil;
 
 public class CardDAO {
@@ -20,96 +19,29 @@ public class CardDAO {
 		this.conn = DBUtil.open();
 	}
 	
-	public int addMyCard(MyCardDTO dto) {
-		
+	public int addCard(CardDTO dto) {
 		try {
 
-			String sql = "insert into tblMyCard (seq, cardNumber, alias, validity, idMember, seqCardInformation) values (seqMyCard.nextVal, ?, ?, ?, ?)";
+			String sql = "insert into tblCardInformation (seq, name, explanation, annualFee, overseasUse, cardCompany, fileLink, seqCardType) values (SEQCARDINFORMATION.NEXTVAL, ?, ?, ?, ?, ?, ?, ?)";
+
 			pstat = conn.prepareStatement(sql);
-			pstat.setString(1, dto.getCardNumber());
-			pstat.setString(2, dto.getAlias());
-			pstat.setString(3, dto.getValidity());
-			pstat.setString(4, dto.getIdMember());
-			pstat.setString(5, dto.getSeqCardInformation());
+			pstat.setString(1, dto.getCiName());
+			pstat.setString(2, dto.getExplanation());
+			pstat.setInt(3, dto.getAnnualFee());
+			pstat.setString(4, dto.getOverseasUse());
+			pstat.setString(5, dto.getCardCompany());
+			pstat.setString(6, dto.getFileLink());
+			pstat.setInt(7, dto.getSeqCardType());
 
 			pstat.executeUpdate();
 
 		} catch (Exception e) {
-			System.out.println("CardDAO.addMyCard");
-			e.printStackTrace();
-		}		
-		
-		return 0;
-	}
-	
-	
-	public int delMyCard(MyCardDTO dto) {
-		
-		try {
-
-			String sql = "delete from tblMyCard where cardNumber = ?";
-
-			pstat = conn.prepareStatement(sql);
-			pstat.setString(1, dto.getCardNumber());
-
-			pstat.executeUpdate();
-
-		} catch (Exception e) {
-			System.out.println("CardDAO.delMyCard");
+			System.out.println("CardDAO.addCard");
 			e.printStackTrace();
 		}
 		return 0;
 	}
 	
-	public int updateMyCard(MyCardDTO dto) {
-		
-		try {
-
-			String sql = "update tblMyCard set alias = ? where cardnumber = ?";
-
-			pstat = conn.prepareStatement(sql);
-			pstat.setString(1, dto.getAlias());
-			pstat.setString(2, dto.getCardNumber());
-
-			pstat.executeUpdate();
-
-		} catch (Exception e) {
-			System.out.println("CardDAO.updateMyCard");
-			e.printStackTrace();
-		}
-		
-		return 0;
-	}
-	
-	public MyCardDTO getMyCard(String cardNumber) {
-		
-		try {
-			
-			String sql = "select * from tblMyCard where cardnumber = ?";
-			
-			pstat = conn.prepareStatement(sql);
-			pstat.setString(1, cardNumber);
-			
-			rs = pstat.executeQuery(sql);
-			
-			if (rs.next()) {
-				
-				MyCardDTO dto = new MyCardDTO();
-				
-				dto.setAlias(rs.getString("alias"));
-				dto.setValidity(rs.getString("validity"));
-				dto.setIdMember(rs.getString("idMember"));
-				dto.setSeqCardInformation(rs.getString("seqCardInformation"));
-				
-				return dto;				
-			}	
-			
-		} catch (Exception e) {
-			e.printStackTrace();
-		}
-		
-		return null;
-	}
 	
 }
